+++ protocols/dcutr/src/protocol/outbound.rs
@@ -23,7 +23,7 @@ use std::io;
 use asynchronous_codec::Framed;
 use futures::prelude::*;
 use futures_timer::Delay;
-use libp2p_core::{Multiaddr, multiaddr::Protocol};
+use libp2p_core::{Multiaddr, MultiaddrExt};
 use libp2p_swarm::Stream;
 use thiserror::Error;
 use web_time::Instant;
@@ -74,7 +74,7 @@ pub(crate) async fn handshake(
         })
         // Filter out relayed addresses.
         .filter(|a| {
-            if a.iter().any(|p| p == Protocol::P2pCircuit) {
+            if a.is_relayed() {
                 tracing::debug!(address=%a, "Dropping relayed address");
                 false
             } else {